Welcome, one and all, to the new style sidebar API.
Nope, there's no documentation yet (but see the poorly titled 
http://www.bofh.org.uk/2006/04/16/write-your-own-typo-sidebar for an
overview of what's new).

The big change is the addition of a bunch of classmethods to the Sidebars::Plugin 
superclass. The new methods are :setting, :display_name and :description.

We already had :display_name and :description, but the new version take an optional
argument and work as declarators as well. Also, :display_name does a good job of
guessing the right display name in many cases.

With the addition of these classmethods, we get to DRY up all those gruesome
boilerplated configure.rhtml files since 'setting' provides enough information
for them to be autogenerated.

Oh, and this change is backwards compatible. You don't have to use the new 
class methods, but you're daft if you don't. You'll still have to get rid of the 
explict form_tag and observe_form bits of really old configure.rhtml files, but 
I can't make us that backward compatible because the old way broke browsers.

If you want to see pretty much all the ways :setting can work, check out 
!XmlController and !FlickrController, which pretty much cover the waterfront.


git-svn-id: http://svn.typosphere.org/typo/trunk@1022 820eb932-12ee-0310-9ca8-eeb645f39767