The beginnings of some serious reworking of the way article publication, notifications, remote pings and (possibly) caching will work. After this patch an article's 'published' flag will never be set before its 'published_at' time.
This is handled by introducing a 'triggers' table which holds a list of actions 
to perform on objects in the future. Every time the rails system gets a request, 
it performs any pending trigger actions. That way, although a message may well
not be 'really' published until some time after its published_at time, it will 
appear as if it was.

Next up, moving various 'on save' actions out of duplicated code in controllers

NB: There's a bug in Rails that means that migration 45 shouldn't work with
mysql or sqlite, hence the utterly gruesome monkey patch in
lib/rails_patch/active_record.rb -- I look 
forward to the next release of Rails when I can get rid of those monstrosities.


git-svn-id: http://svn.typosphere.org/typo/trunk@1042 820eb932-12ee-0310-9ca8-eeb645f39767