Okay, this is the big, scary patch.
Content is now using a state object to handle transitions between 
'new', 'draft', 'just_published', 'published' and 'publication_pending'
states. The interesting states in that lot are:

  * just_published -- this is a transient state, nothing gets fetched 
    from the object store in this state. content is marked as just 
    published so that various things that should only happen when an 
    article is just published (notifications, pings, etc) only happen
    once.

  * publication_pending -- this is when content is due to be published
    in the future. It generally means there's a trigger hanging about
    in the triggers table waiting to publish the article.

At the moment the publication state isn't stored in the database, but derived 
from the 'published' flag and the 'published_at' field. I think we can
survive like this for a while, but we may end up getting rid of the 'published'
flag and replacing it with a publication_state field holding a simple
memento that we can use to reconstruct the state object. (We can't just
use slaird's {{{acts_as_state}}} plugin because that doesn't appear
to use 'real' state objects that we can hang behaviour off, and I've 
switched us to using the state pattern precisely so I can hang behaviour
off the states -- check out {{{ContentState::JustPublished}}} for an
example of the kind of thing the state pattern lets you do.)

Also, I've switched all the views I could find over to using 
{{{published_at}}} rather than {{{created_at}}}; if you're a theme author
you might like to do the same.

git-svn-id: http://svn.typosphere.org/typo/trunk@1047 820eb932-12ee-0310-9ca8-eeb645f39767