The RESTful Typo revolution continues apace. This patch moves most feeds from /xml/... to simply being at their 'equivalent' url with a .atom or .rss extension as appropriate. So /xml/atom/feed.xml becomes /articles.atom and (slightly more cunningly), /xml/atom/article/1/feed.xml becomes /articles/yyyy/mm/dd/slug.atom.
XmlController isn't going away in the forseeable future, but it'll be moving
towards being a pure 'redirect' controller for historic URLs. Once it *is*
a pure redirector we'll look to moving it out into the plugins directory
for use by folks who are upgrading and want to keep their historic feed 
URLs catching the occasional dumb feed reader, but new installs simply 
won't need it.

git-svn-id: http://svn.typosphere.org/typo/trunk@1467 820eb932-12ee-0310-9ca8-eeb645f39767