Adds a "setup" in the environment.rb, running migrations if the table session doesn't exist.
I wanted to create a cool setup page allowing the user a 1 click install like on many blogwares, catching an exception in the application.rb. It seems that rails tries to select from the sessions table before my application can do anything and throw an exception resulting in a 500 error. So, the only place to put that code snippet was in application.rb. Irk !

git-svn-id: http://svn.typosphere.org/typo/branches/rails_2@1589 820eb932-12ee-0310-9ca8-eeb645f39767