Giving tests and fixtures some love.
Moving from explicit ids throughout the fixtures and many of the tests to taking
advantage of the new fixture capabilities in Rails 2. 

We're now loading _all_ the fixtures from test_helper, which should hopefully
reduce problems we've been experiencing with test ordering, which is nice.

Eventually, I intend to move to using pre_loaded_fixtures - loading the fixtures
up as part of the initial rake task, and then letting transactional rollback
stuff deal with the rest.

git-svn-id: http://svn.typosphere.org/typo/trunk@1620 820eb932-12ee-0310-9ca8-eeb645f39767